+++ R/zzz.R
@@ -7,9 +7,9 @@
 # without having knitr as a dependency for this package
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 register_s3_method <- function(pkg, generic, class, fun = NULL) {
-  stopifnot(is.character(pkg), length(pkg) == 1)
+  stopifnot(is.character(pkg)    , length(pkg) == 1)
   stopifnot(is.character(generic), length(generic) == 1)
-  stopifnot(is.character(class), length(class) == 1)
+  stopifnot(is.character(class)  , length(class) == 1)
 
   if (is.null(fun)) {
     fun <- get(paste0(generic, ".", class), envir = parent.frame())